LF-5277 Fix animalOptionsSelector and move the file to store/selectors/
